+++ src/com/teamkn/activity/chat/SelectChatMemberActivity.java
@@ -2,15 +2,12 @@ package com.teamkn.activity.chat;
 
 import java.util.ArrayList;
 import java.util.List;
-
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
-import android.view.View.OnClickListener;
 import android.widget.AdapterView;
 import android.widget.AdapterView.OnItemClickListener;
 import android.widget.ListView;
-
 import com.teamkn.R;
 import com.teamkn.Logic.HttpApi;
 import com.teamkn.base.activity.TeamknBaseActivity;
@@ -19,7 +16,6 @@ import com.teamkn.base.utils.BaseUtils;
 import com.teamkn.model.Contact;
 import com.teamkn.model.database.ChatDBHelper;
 import com.teamkn.model.database.ContactDBHelper;
-import com.teamkn.service.SynNoteService;
 import com.teamkn.widget.adapter.SelectChatMemberListAdapter;
 
 public class SelectChatMemberActivity extends TeamknBaseActivity {
@@ -66,12 +62,12 @@ public class SelectChatMemberActivity extends TeamknBaseActivity {
   public void click_submit_select_chat_member(View view){
     if(select_chat_member_ids.size() == 0){return;}
     
-    new TeamknAsyncTask<Void,Void,Long>(this,"正在创建") {
+    new TeamknAsyncTask<Void,Void,Integer>(this,"正在创建") {
 
       @Override
-      public Long do_in_background(Void... params) throws Exception {
+      public Integer do_in_background(Void... params) throws Exception {
         List<Integer> user_list = select_chat_member_ids;
-        long client_chat_id = ChatDBHelper.create(user_list);
+        int client_chat_id = ChatDBHelper.create(user_list);
         if(BaseUtils.is_wifi_active(SelectChatMemberActivity.this)){
           HttpApi.Chat.create(client_chat_id,user_list);
         }
@@ -79,9 +75,9 @@ public class SelectChatMemberActivity extends TeamknBaseActivity {
       }
 
       @Override
-      public void on_success(Long client_chat_id) {
+      public void on_success(Integer client_chat_id) {
         Intent intent = new Intent(SelectChatMemberActivity.this,ChatActivity.class);
-        intent.putExtra(ChatActivity.Extra.CLIENT_CHAT_ID, (long)client_chat_id);
+        intent.putExtra(ChatActivity.Extra.CLIENT_CHAT_ID, (int)client_chat_id);
         startActivity(intent);
         SelectChatMemberActivity.this.finish();
       }
